+++ src/main/java/redstonedubstep/mods/vanishmod/mixin/MixinServerGamePacketListenerImpl.java
@@ -9,6 +9,10 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
+import net.minecraft.network.PacketSendListener;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.contents.TranslatableContents;
 import net.minecraft.network.protocol.Packet;
 import net.minecraft.network.protocol.game.ClientboundBlockEventPacket;
 import net.minecraft.network.protocol.game.ClientboundLevelEventPacket;
@@ -17,13 +21,16 @@ import net.minecraft.network.protocol.game.ClientboundPlayerInfoPacket.Action;
 import net.minecraft.network.protocol.game.ClientboundPlayerInfoPacket.PlayerUpdate;
 import net.minecraft.network.protocol.game.ClientboundSoundEntityPacket;
 import net.minecraft.network.protocol.game.ClientboundSoundPacket;
+import net.minecraft.network.protocol.game.ClientboundSystemChatPacket;
 import net.minecraft.network.protocol.game.ClientboundTakeItemEntityPacket;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.network.ServerGamePacketListenerImpl;
 import net.minecraft.world.phys.Vec3;
+import net.minecraftforge.server.ServerLifecycleHooks;
 import redstonedubstep.mods.vanishmod.VanishConfig;
 import redstonedubstep.mods.vanishmod.VanishUtil;
+import redstonedubstep.mods.vanishmod.misc.FieldHolder;
 import redstonedubstep.mods.vanishmod.misc.SoundSuppressionHelper;
 
 @Mixin(ServerGamePacketListenerImpl.class)
@@ -40,12 +47,14 @@ public class MixinServerGamePacketListenerImpl {
 	@Inject(method = "send(Lnet/minecraft/network/protocol/Packet;)V", at = @At("HEAD"), cancellable = true)
 	private void onSendPacket(Packet<?> packet, CallbackInfo callbackInfo) {
 		if (packet instanceof ClientboundPlayerInfoPacket infoPacket && infoPacket.getAction() != Action.REMOVE_PLAYER) {
-			List<PlayerUpdate> vanishedPacketEntries = infoPacket.getEntries().stream().filter(p -> VanishUtil.isVanished(server.getPlayerList().getPlayer(p.getProfile().getId()), player)).toList();
+			List<PlayerUpdate> filteredPacketEntries = infoPacket.getEntries().stream().filter(p -> !VanishUtil.isVanished(server.getPlayerList().getPlayer(p.getProfile().getId()), player)).toList();
 
-			if (vanishedPacketEntries.equals(infoPacket.getEntries()))
+			if (filteredPacketEntries.isEmpty())
 				callbackInfo.cancel();
-			else if (!vanishedPacketEntries.isEmpty())
-				infoPacket.getEntries().removeAll(vanishedPacketEntries);
+			else if (!filteredPacketEntries.equals(infoPacket.getEntries())) {
+				infoPacket.getEntries().clear();
+				infoPacket.getEntries().addAll(filteredPacketEntries);
+			}
 		}
 		else if (packet instanceof ClientboundTakeItemEntityPacket pickupPacket && VanishUtil.isVanished(player.level.getEntity(pickupPacket.getPlayerId()), player))
 			callbackInfo.cancel();
@@ -60,4 +69,24 @@ public class MixinServerGamePacketListenerImpl {
 				callbackInfo.cancel();
 		}
 	}
+
+	//Prevent join, leave, death and advancement messages of vanished players from being broadcast
+	@Inject(method = "send(Lnet/minecraft/network/protocol/Packet;Lnet/minecraft/network/PacketSendListener;)V", at = @At("HEAD"), cancellable = true)
+	private void onSendPacket(Packet<?> packet, PacketSendListener listener, CallbackInfo callbackInfo) {
+		if (packet instanceof ClientboundSystemChatPacket chatPacket && chatPacket.content() instanceof MutableComponent component && component.getContents() instanceof TranslatableContents content) {
+			if (content.getKey().startsWith("multiplayer.player.joined") && VanishUtil.isVanished(FieldHolder.joiningPlayer)) {
+				FieldHolder.joiningPlayer = null;
+				callbackInfo.cancel();
+			}
+			else if (content.getKey().startsWith("multiplayer.player.left") || content.getKey().startsWith("death.") || content.getKey().startsWith("chat.type.advancement")) {
+				if (content.getArgs()[0] instanceof Component playerName) {
+					for (ServerPlayer sender : ServerLifecycleHooks.getCurrentServer().getPlayerList().getPlayers()) {
+						if (sender.getDisplayName().getString().equals(playerName.getString()) && VanishUtil.isVanished(sender, player))
+							callbackInfo.cancel();
+					}
+				}
+			}
+		}
+	}
+
 }