+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - Fixed vanished players not hearing the sounds they produce
 - Fixed vanished players not seeing an item pickup animation when they collect an item
+- Fixed vanished players not receiving join/leave/death messages of other vanished players if they are allowed to see them
 - Fixed chests and ender chests visually opening when being opened by a vanished player
 - Fixed some entities like passive mobs or pufferfish detecting vanished players
 - The minimum required Forge version is now 43.1.8